requirements.txt
aiofiles
aiohttp
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
lyricsgenius
motor
ntgcalls
pillow
psutil
pykeyboard
py-tgcalls
git+https://github.com/KurimuzonAkuma/pyrogram@dev
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
tgcrypto
uvloop
youtube-search
youtube-search-python
yt-dlp
